acceptance/serverless-db-tracing/requirements.txt
git+git://github.com/iopipe/iopipe-python.git@master#egg=iopipe
mysqlclient
psycopg2-binary
